+++ solve.c
@@ -6,14 +6,13 @@
 /*   By: mguerrea <mguerrea@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/10/16 23:30:41 by lbenard           #+#    #+#             */
-/*   Updated: 2018/11/11 15:00:59 by mguerrea         ###   ########.fr       */
+/*   Updated: 2018/11/12 15:49:41 by lbenard          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "functions.h"
-#include <stdio.h>
 
-int	check_place(t_grid *grid, t_pos *pos, t_tetrimino *tetrimino)
+int		check_place(t_grid *grid, t_pos *pos, t_tetrimino *tetrimino)
 {
 	t_list	*blocks;
 	t_pos	origin;
@@ -28,7 +27,7 @@ int	check_place(t_grid *grid, t_pos *pos, t_tetrimino *tetrimino)
 		origin = *pos;
 		origin.x += ((t_pos*)blocks->content)->x;
 		origin.y += ((t_pos*)blocks->content)->y;
-		if ((origin.x >= grid->width || origin.y >= grid->width) ||
+		if ((origin.x >= (int)grid->width || origin.y >= (int)grid->width) ||
 			*get_char_at(&origin, grid) != '.')
 			return (0);
 		blocks = blocks->next;
@@ -36,7 +35,7 @@ int	check_place(t_grid *grid, t_pos *pos, t_tetrimino *tetrimino)
 	return (1);
 }
 
-void    place_tetrimino(t_grid *grid, t_tetrimino *tetrimino, t_pos *pos)
+void	place_tetrimino(t_grid *grid, t_tetrimino *tetrimino, t_pos *pos)
 {
 	t_list	*blocks;
 	t_pos	origin;
@@ -76,15 +75,14 @@ void	delete_tetrimino(t_grid *grid, t_tetrimino *tetrimino, t_pos *pos)
 	}
 }
 
-int    solve(t_grid *grid, t_pos pos, t_tetrimino *list)
+int		solve(t_grid *grid, t_pos pos, t_tetrimino *list)
 {
-	while (!(pos.x == grid->width - 1 && pos.y == grid->width -1))
+	while (!(pos.x == (int)grid->width - 1 && pos.y == (int)grid->width - 1))
 	{
 		if (check_place(grid, &pos, list))
 		{
 			place_tetrimino(grid, list, &pos);
 			set_pos(&pos, 0, 0);
-	//		printf("%s\n", grid->str);
 			if ((list + 1)->letter)
 			{
 				if (solve(grid, pos, list + 1))
